Exception Handling

+++ mpcommerce/src/main/java/com/devsuperior/mpcommerce/Services/ProductService.java
@@ -1,13 +1,16 @@
 package com.devsuperior.mpcommerce.Services;
 
+import com.devsuperior.mpcommerce.Services.exceptions.DatabaseException;
 import com.devsuperior.mpcommerce.Services.exceptions.ResourceNotFoundException;
 import com.devsuperior.mpcommerce.dto.ProductDTO;
 import com.devsuperior.mpcommerce.entities.Product;
 import com.devsuperior.mpcommerce.repositories.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import java.util.Optional;
 
@@ -50,12 +53,17 @@ public class ProductService {
         return new ProductDTO(entity);
     }
 
-
-    @Transactional
-    public void delete(Long id){
-
-        repository.deleteById(id);
-
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void CopyDtoToEntity(ProductDTO dto, Product entity) {
@@ -66,6 +74,8 @@ public class ProductService {
         entity.setImgUrl(dto.getImgUrl());
 
     }
+
+
 
 
 }
